Move the folders up
