+++ console-app-device-code/requirements.txt
