+++ protect-function-app/.gitignore
