+++ protect-function-app/greeting/function.json
