+++ protect-function-app/host.json
